+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-10-17
+### Added
+- The `Container::getCacheFile()` now accepts an optional callable parameter for encoding values.
+
 ### Changed
+- The container now stores cached information only in a single property
 - Improved tests and some failure conditions with mutation testing
 
 ## [0.2.1] - 2018-07-18
@@ -26,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial development release
 
-[Unreleased]: https://github.com/simply-framework/container/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/simply-framework/container/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/simply-framework/container/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/simply-framework/container/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/simply-framework/container/compare/v0.1.0...v0.2.0